add "struct coord" type, flatten input event arglists
makes working with all those coordinates everywhere easier. like
the GLSL vec2, somehow.

now that the number of fields in the event struct has effectively
halved, the event handlers can use flat lists that are as clear as
the struct was before. just that input.c is waaay more readable now