thumbnail: add vertical border and compute total height
Horizontal border was killed for good but there should be a vertical
border. Or so I think. (Just a matter of time until I replace that
with spacing as well)

Also draws a line at the bottom to show where the generated image ends.
(Will be removed once scrolling into empty space is no longer allowed.)