thumbnail: update thumbnails in background, not only on X events
Maintains a flag that states if there are thumbnails to be updated,
in which case the select loop continues to generate more thumbnails
instead of waiting for user input.

Obscure bugs with Xlib (see that big comment in main.c) have probably
been fixed.

Also, doing a redraw now actually saves that a redraw has happened and
prevents another useless one from happening.