no longer generate thumbnails exactly the needed size but 128/256/512
Instead generate thumbnails that fit into 128x128, 256x256 or 512x512
(depending on the size wanted).

The thumbnails will be rescaled to the real size on rendering. There's
no caching of the "odd" sizes like 200x150. Benchmarks have shown that
this has a 1% speed impact on rendering which is irrelevant compared to
the time needed to actually read the image. But it saves us from
refreshing all thumbnails on every resize, which is probably better.

Currently thumbnails larger than 512x512 will not be generated, so for
these sizes which shouldn't be called thumbnails the previews will be
blurry. This is probably acceptable.