common: x11_window -> x11.window
Really no reason to keep it separate.

Well, there is. Everything but x11_window is not referring to a window,
so .window is misplaced if for some reason more than one window is used
(menus?!).

But for now that's not a problem and it can be very easily reversed if
necessary.